Minimize compiling warnings in ltjs_lib_crypt_mgr

+++ libs/cryptmgr/cryptmgr.cpp
@@ -40,7 +40,7 @@ void CCryptMgr::Encrypt(std::istream& is, std::ostream& os)
 	{
 		memset(buf, 0, 8);
 		is.read(buf, 8);
-		n = is.gcount();
+		n = static_cast<int>(is.gcount());
 		Blowfish_encipher((UWORD_32bits*)buf, (UWORD_32bits*)&buf[4]);
 		os.write(buf, 8);
 	}
@@ -60,7 +60,7 @@ void CCryptMgr::Decrypt(std::istream& is, std::ostream& os)
 	while (!is.eof())
 	{
 		is.read(buf, 8);
-		n = is.gcount();
+		n = static_cast<int>(is.gcount());
 		if (n == 1)
 			n = (int)buf[0];
 		if (!bFirstTime)
